+++ src/internal/FFTConvolver.h
@@ -16,7 +16,7 @@ class FFTConvolver
 {
 public:
     // fftSize must be a power of two
-    FFTConvolver(size_t fftSize);
+    FFTConvolver(int fftSize);
 
     // For now, with multiple calls to Process(), framesToProcess MUST add up EXACTLY to fftSize / 2
     //
@@ -25,11 +25,11 @@ public:
     // The input to output latency is equal to fftSize / 2
     //
     // Processing in-place is allowed...
-    void process(FFTFrame * fftKernel, const float * sourceP, float * destP, size_t framesToProcess);
+    void process(FFTFrame * fftKernel, const float * sourceP, float * destP, int framesToProcess);
 
     void reset();
 
-    size_t fftSize() const { return m_frame.fftSize(); }
+    int fftSize() const { return m_frame.fftSize(); }
 
 private:
     FFTFrame m_frame;
